requirements.txt
Flask==2.3.3
azure-storage-blob==12.14.0
